fix(avatar): display default icon in case the picture src is null/not present
