+++ src/components/freelancer/FreelancerStrip/index.tsx
@@ -1,13 +1,14 @@
 import Chip from '@material-ui/core/Chip'
-import Image from 'next/image'
 import React from 'react'
+
+import Avatar from '../../common/Avatar'
 
 const FreelancerStrip = ({ freelancer, rate, updateFreelancerIndex, currency, index }) => {
   const { picture, name, title, skills } = freelancer
   return (
     <button type="button" onClick={() => updateFreelancerIndex(index)} className="flex p-3 m-4 border-2 border-solid">
       <div className="flex flex-col items-center">
-        {picture ? <Image height={68} width={68} src={picture} className="rounded-full" /> : null}
+        <Avatar src={picture} tailwindSizeClass="h-16 w-16" />
         <p>
           {currency}
           {rate}
